requirements.txt
# Neural Language Steering - Complete Requirements
# For Llama 3.1 8B implementation on A100 hardware

# CRITICAL: Fix Keras compatibility
tf-keras>=2.15.0

# Core PyTorch and CUDA
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
--extra-index-url https://download.pytorch.org/whl/cu118

# Transformers and HuggingFace ecosystem
transformers>=4.36.0
accelerate>=0.24.0
datasets>=2.14.0
tokenizers>=0.15.0
huggingface-hub>=0.19.0

# Deep learning optimization
flash-attn>=2.3.0
xformers>=0.0.22
triton>=2.1.0
deepspeed>=0.12.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Language processing
fasttext>=0.9.2
langdetect>=1.0.9
sacremoses>=0.0.53
indic-nlp-library>=0.81

# Machine learning utilities
scikit-learn>=1.3.0
tqdm>=4.66.0
wandb>=0.16.0
tensorboard>=2.15.0

# Data processing
pyarrow>=14.0.0
jsonlines>=4.0.0
regex>=2023.10.3

# Web interfaces and visualization
streamlit>=1.28.0
gradio>=4.0.0
jupyter>=1.0.0
ipywidgets>=8.1.0

# System monitoring
psutil>=5.9.0
gpustat>=1.1.0
nvidia-ml-py>=12.535.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: For specific datasets
indiccorp>=0.0.1
ai4bharat-transliteration>=1.0.0

# Audio/speech (if needed for multimodal)
librosa>=0.10.0
soundfile>=0.12.0

# Image processing (if needed)
Pillow>=10.0.0
opencv-python>=4.8.0

# Additional utilities
click>=8.1.0
rich>=13.6.0
typer>=0.9.0
pydantic>=2.7.0

# Enhanced language detection and judging
steering-vectors>=0.2.0

# For SSH deployment
paramiko>=3.3.0
fabric>=3.2.0
invoke>=2.2.0

# Memory profiling
memory-profiler>=0.61.0
pympler>=0.9.0

# Async processing
aiohttp>=3.9.0
uvloop>=0.19.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0
python-dotenv>=1.0.0

# Optional: LLM-as-judge (Gemini)
google-generativeai>=0.7.2

# Note: If you install Torch from a CUDA-specific index URL manually (setup.txt step 5),
# you may want to install requirements with `--no-deps` or install Torch first,
# then `pip install -r requirements.txt` to avoid index conflicts.

# Optional kernel accelerators (auto-registered if installed)
lookahead-keys-attention>=0.1.0
liger-kernel>=0.5.10
